# apidoc/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "The OpenFlorian REST API enables management of the server via REST",
    "version" : "v1",
    "title" : "Openflorian REST API",
    "termsOfService" : "http://bitbucket.org/ceth/openflorian",
    "contact" : {
      "name" : "Bastian Kraus",
      "url" : "http://about.me/bastian.kraus",
      "email" : "bofh@cpotr.de"
    },
    "license" : {
      "name" : "GPL v3",
      "url" : "hhttp://www.gnu.de/documents/gpl.en.html"
    }
  },
  "host" : "of-server",
  "basePath" : "/openflorian/rest/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/alarm" : {
      "post" : {
        "tags" : [ "alarm" ],
        "summary" : "Incurre alarm",
        "description" : "",
        "operationId" : "incurre",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "operationId",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
...
}